+++ rustfst/src/semirings/probability_weight.rs
@@ -2,14 +2,17 @@ use std::borrow::Borrow;
 use std::f32;
 use std::hash::{Hash, Hasher};
 
-use failure::Fallible;
-use ordered_float::OrderedFloat;
-
+use crate::parsers::bin_fst::utils_serialization::write_bin_f32;
 use crate::semirings::{
-    CompleteSemiring, DivideType, ReverseBack, Semiring, SemiringProperties, StarSemiring,
-    WeaklyDivisibleSemiring, WeightQuantize,
+    CompleteSemiring, DivideType, ReverseBack, Semiring, SemiringProperties, SerializableSemiring,
+    StarSemiring, WeaklyDivisibleSemiring, WeightQuantize,
 };
 use crate::KDELTA;
+use failure::Fallible;
+use nom::number::complete::{float, le_f32};
+use nom::IResult;
+use ordered_float::OrderedFloat;
+use std::io::Write;
 
 /// Probability semiring: (x, +, 0.0, 1.0).
 #[derive(Clone, Debug, PartialOrd, Default, Copy, Eq)]
@@ -87,6 +90,26 @@ display_semiring!(ProbabilityWeight);
 
 impl CompleteSemiring for ProbabilityWeight {}
 
+impl SerializableSemiring for ProbabilityWeight {
+    fn weight_type() -> String {
+        "probability".to_string()
+    }
+
+    fn parse_binary(i: &[u8]) -> IResult<&[u8], Self> {
+        let (i, weight) = le_f32(i)?;
+        Ok((i, Self::new(weight)))
+    }
+
+    fn write_binary<F: Write>(&self, file: &mut F) -> Fallible<()> {
+        write_bin_f32(file, *self.value())
+    }
+
+    fn parse_text(i: &str) -> IResult<&str, Self> {
+        let (i, f) = float(i)?;
+        Ok((i, Self::new(f)))
+    }
+}
+
 impl StarSemiring for ProbabilityWeight {
     fn closure(&self) -> Self {
         Self::new(1.0 / (1.0 - self.value.0))
@@ -107,6 +130,12 @@ impl WeaklyDivisibleSemiring for ProbabilityWeight {
 impl_quantize_f32!(ProbabilityWeight);
 
 partial_eq_and_hash_f32!(ProbabilityWeight);
+
+test_semiring_serializable!(
+    tests_probability_weight_serializable,
+    ProbabilityWeight,
+    ProbabilityWeight::one() ProbabilityWeight::zero() ProbabilityWeight::new(0.3) ProbabilityWeight::new(0.5) ProbabilityWeight::new(0.0) ProbabilityWeight::new(1.0)
+);
 
 impl Into<ProbabilityWeight> for f32 {
     fn into(self) -> ProbabilityWeight {