+++ rustfst/src/tests_openfst/io/const_fst_text_serialization.rs
@@ -3,8 +3,9 @@ use tempfile::tempdir;
 
 use crate::algorithms::fst_convert_from_ref;
 use crate::fst_impls::{ConstFst, VectorFst};
-use crate::fst_traits::SerializableFst;
+use crate::fst_traits::{Fst, SerializableFst};
 use crate::semirings::SerializableSemiring;
+use crate::tests_openfst::io::generate_symbol_table;
 use crate::tests_openfst::FstTestData;
 
 pub fn test_const_fst_text_serialization<W>(test_data: &FstTestData<VectorFst<W>>) -> Fallible<()>
@@ -39,6 +40,42 @@ where
             test_data.raw,
             deserialized_fst_vec,
             "Serializer VectorFst Text"
+        )
+    );
+    Ok(())
+}
+
+pub fn test_const_fst_text_serialization_with_symt<W>(
+    test_data: &FstTestData<VectorFst<W>>,
+) -> Fallible<()>
+where
+    W: SerializableSemiring + 'static,
+{
+    let dir = tempdir()?;
+
+    let mut raw_const_with_symt: ConstFst<_> = test_data.raw.clone().into();
+    let (input_symt, output_symt) = generate_symbol_table("test", &raw_const_with_symt);
+    raw_const_with_symt.set_input_symbols(input_symt);
+    raw_const_with_symt.set_output_symbols(output_symt);
+
+    let path_fst_serialized = dir.path().join("raw_const.fst");
+    raw_const_with_symt.write_text(&path_fst_serialized)?;
+
+    let deserialized_fst = ConstFst::<W>::read_text(&path_fst_serialized)?;
+
+    // Text serialization doesn't include the symbol table.
+    let mut raw_const_without_symt = raw_const_with_symt;
+    raw_const_without_symt.unset_input_symbols();
+    raw_const_without_symt.unset_output_symbols();
+
+    assert_eq!(
+        raw_const_without_symt,
+        deserialized_fst,
+        "{}",
+        error_message_fst!(
+            raw_const_without_symt,
+            deserialized_fst,
+            "Serializer ConstFst Text with Generated Symbol Table"
         )
     );
     Ok(())