+++ rustfst/src/tests_openfst/mod.rs
@@ -30,8 +30,12 @@ use crate::tests_openfst::algorithms::gallic_encode_decode::GallicTestData;
 use crate::tests_openfst::io::const_fst_bin_deserializer::{
     test_const_fst_aligned_bin_deserializer, test_const_fst_bin_deserializer,
 };
-use crate::tests_openfst::io::const_fst_bin_serializer::test_const_fst_bin_serializer;
-use crate::tests_openfst::io::const_fst_text_serialization::test_const_fst_text_serialization;
+use crate::tests_openfst::io::const_fst_bin_serializer::{
+    test_const_fst_bin_serializer, test_const_fst_bin_serializer_with_symt,
+};
+use crate::tests_openfst::io::const_fst_text_serialization::{
+    test_const_fst_text_serialization, test_const_fst_text_serialization_with_symt,
+};
 
 use self::algorithms::{
     arc_map::{
@@ -65,9 +69,6 @@ use self::fst_impls::const_fst::test_const_fst_convert_convert;
 use self::fst_impls::test_fst_into_iterator::{
     test_fst_into_iterator_const, test_fst_into_iterator_vector,
 };
-use self::io::vector_fst_bin_deserializer::test_vector_fst_bin_deserializer;
-use self::io::vector_fst_bin_serializer::test_vector_fst_bin_serializer;
-use self::io::vector_fst_text_serialization::test_vector_fst_text_serialization;
 use self::misc::test_del_all_states;
 use crate::fst_traits::SerializableFst;
 use crate::tests_openfst::algorithms::closure::{
@@ -78,7 +79,14 @@ use crate::tests_openfst::algorithms::concat::{
     test_concat, test_concat_dynamic, ConcatOperationResult, ConcatTestData,
 };
 use crate::tests_openfst::algorithms::union::{test_union, test_union_dynamic};
+use crate::tests_openfst::io::vector_fst_bin_deserializer::test_vector_fst_bin_deserializer;
 use crate::tests_openfst::io::vector_fst_bin_deserializer::test_vector_fst_bin_with_symt_deserializer;
+use crate::tests_openfst::io::vector_fst_bin_serializer::{
+    test_vector_fst_bin_serializer, test_vector_fst_bin_serializer_with_symt,
+};
+use crate::tests_openfst::io::vector_fst_text_serialization::{
+    test_vector_fst_text_serialization, test_vector_fst_text_serialization_with_symt,
+};
 
 #[macro_use]
 mod macros;
@@ -606,8 +614,20 @@ macro_rules! test_fst {
             }
 
             #[test]
+            fn test_vector_fst_text_serialization_with_symt_openfst() -> Fallible<()> {
+                do_run!(test_vector_fst_text_serialization_with_symt, $fst_name);
+                Ok(())
+            }
+
+            #[test]
             fn test_vector_fst_bin_serializer_openfst() -> Fallible<()> {
                 do_run!(test_vector_fst_bin_serializer, $fst_name);
+                Ok(())
+            }
+
+            #[test]
+            fn test_vector_fst_bin_serializer_with_symt_openfst() -> Fallible<()> {
+                do_run!(test_vector_fst_bin_serializer_with_symt, $fst_name);
                 Ok(())
             }
 
@@ -654,8 +674,20 @@ macro_rules! test_fst {
             }
 
             #[test]
+            fn test_const_fst_bin_serializer_with_symt_openfst() -> Fallible<()> {
+                do_run!(test_const_fst_bin_serializer_with_symt, $fst_name);
+                Ok(())
+            }
+
+            #[test]
             fn test_const_fst_text_serialization_openfst() -> Fallible<()> {
                 do_run!(test_const_fst_text_serialization, $fst_name);
+                Ok(())
+            }
+
+            #[test]
+            fn test_const_fst_text_serialization_with_symt_openfst() -> Fallible<()> {
+                do_run!(test_const_fst_text_serialization_with_symt, $fst_name);
                 Ok(())
             }